requirements.txt
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
dnspython==1.16.0
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.19.0
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
pymongo==3.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
RapidFuzz==3.13.0
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.4
